#5 - breaking up model layer into different files
