+++ model/MarvelCharacter.ts
@@ -1,0 +1,9 @@
+export interface IMarvelCharacter {
+  name: string;
+  image: string;
+}
+
+export const MarvelCharacter = ({ name = '', image = '' } = {}): IMarvelCharacter => ({
+  name,
+  image
+});
